Annotate Each Provided Differentialy Expressed Region (DER) with Its Closest Annotation(s)

annotateNearest() relies on user-specified annotations in GFF or GTF format to annotate the DERs found during the differential expression analysis. Seven classes of labels translate the relative positions of the stranded DER and its closest annotation(s):

antisense ;

upstream ;

downstream ;

inside ;

overlapping 3’ ;

overlapping 5’ ;

overlapping 5’, 3’.

For the non-stranded DERs, the two classes of labels are: overlapping and non-overlapping.

Usage

annotateNearest(
  features,
  annotations,
  outputDirectory = NULL,
  annotationsType = NULL,
  orderBy = "start",
  select = NULL,
  verbose = TRUE
)

Arguments

features

A GenomicRanges object augmented with the DER column, which indicates if the region is differentially expressed, as returned by DiffSegR::dea().

annotations

A String. Path to the annotation file (in GFF or GTF format).

outputDirectory

A String. Path to the output directory.

annotationsType

A vector of String. Keep only matching annotations.

orderBy

A String. The name of the column on which returned annotations are sorted.

select

A vector of String. Keep only columns matching these names in the returned Data.frame. By default all columns from mcols(dds) and mcols(annotations) are kept.

verbose

A Boolean. Should all the operations performed be displayed ?

Value

Annotated differentialy expressed regions as Data.frame object.