Differential expression analysis (DEA)

dea takes as input the recently built SummarizedExperiment object and calls DESeq2 to test for each line, i.e. each region, the difference in expression between the two compared biological conditions. The user can choose to control all the resulting p-values or only a subset of interest. In the first case a Benjamini–Hochberg procedure is used to control the False Discovery Rate (FDR). In the second case, the user specifies selection criteria for the regions, e.g. a threshold on the absolute log2-FC per-regions. Then, a post-hoc procedure is used to control the joint Family wise error Rate (jFWER) on the selected subset of regions.

Usage

dea(
  data,
  SExp,
  design = ~condition,
  sizeFactors = NA,
  significanceLevel = 0.05,
  predicate = NULL,
  postHoc_significanceLevel = 0.05,
  postHoc_tdpLowerBound = 0.95,
  orderBy = "pvalue",
  verbose = FALSE,
  dichotomicSearch = FALSE
)

Arguments

data

The List object returned by loadData().

SExp

The SummarizedExperiment object returned by segmentation().

design

A formula object or a Matrix object. Passed on to DESeq2::DESeqDataSet.

sizeFactors

A vector of Double. Sample-specific size factors.

significanceLevel

A Double. The significance cutoff on q-values.

predicate

A predicate function that returns a single TRUE or FALSE if the region on which it is applied meets the conditions defined in the predicate. The criteria used in the predicate have to be defined for each regions in mcols(SExp).

postHoc_significanceLevel

A Double. The significance level of the test procedure (see sanssouci::posthocBySimes).

postHoc_tdpLowerBound

A Double. The minimum true positive proportion on the returned set of rejected regions.

orderBy

A String. The parameter on which regions are sorted before the post-hoc procedure.

verbose

A Boolean. Should all the operations performed be displayed (only post-hoc) ?

dichotomicSearch

A Boolean. Use it to speed up the post-hoc procedure.

Value

ADESeqDataSet object.

Examples

# Create a working directory for running the example.
working_directory <- "./DIFFSEGR_TEST"
dir.create(working_directory)

# Save sample information in a text file.
sample_info <- data.frame(
  sample    = c("pnp1_1_1", "pnp1_1_2", "wt_1", "wt_2"),
  condition = rep(c("pnp1_1", "wt"), each = 2),
  replicate = rep(1:2,2),
  bam       = sapply(
    c("pnp1_1_1_ChrC_71950_78500.bam", 
      "pnp1_1_2_ChrC_71950_78500.bam",
      "wt_1_ChrC_71950_78500.bam",
      "wt_2_ChrC_71950_78500.bam"
     ),
     function(bam) system.file("extdata", bam, package = "DiffSegR")
  ),
  coverage  = file.path(
    working_directory,
    paste0(c("pnp1_1_1", "pnp1_1_2", "wt_1", "wt_2"), ".rds")
  )
)
write.table(
  sample_info, 
  file.path(working_directory, "sample_info.txt")
)

# Build coverages and log2-FC per-base.
data <- loadData(
  sampleInfo         = file.path(working_directory,"sample_info.txt"),
  locus        = list(
    seqid      = "ChrC", 
    chromStart = 71950, 
    chromEnd   = 78500
  ),
  referenceCondition = "wt",
  stranded           = TRUE,
  fromBam            = TRUE,
  nbThreads          = 1
)
#> building coverage profiles ...
#> 
#>         ==========     _____ _    _ ____  _____  ______          _____  
#>         =====         / ____| |  | |  _ \|  __ \|  ____|   /\   |  __ \ 
#>           =====      | (___ | |  | | |_) | |__) | |__     /  \  | |  | |
#>             ====      \___ \| |  | |  _ <|  _  /|  __|   / /\ \ | |  | |
#>               ====    ____) | |__| | |_) | | \ \| |____ / ____ \| |__| |
#>         ==========   |_____/ \____/|____/|_|  \_\______/_/    \_\_____/
#>        Rsubread 2.10.4
#> 
#> //========================== featureCounts setting ===========================\\
#> ||                                                                            ||
#> ||             Input files : 4 BAM files                                      ||
#> ||                                                                            ||
#> ||                           pnp1_1_1_ChrC_71950_78500.bam                    ||
#> ||                           pnp1_1_2_ChrC_71950_78500.bam                    ||
#> ||                           wt_1_ChrC_71950_78500.bam                        ||
#> ||                           wt_2_ChrC_71950_78500.bam                        ||
#> ||                                                                            ||
#> ||              Paired-end : no                                               ||
#> ||        Count read pairs : no                                               ||
#> ||              Annotation : R data.frame                                     ||
#> ||      Dir for temp files : .                                                ||
#> ||                 Threads : 1                                                ||
#> ||                   Level : meta-feature level                               ||
#> ||      Multimapping reads : counted                                          ||
#> || Multi-overlapping reads : not counted                                      ||
#> ||   Min overlapping bases : 1                                                ||
#> ||          Read reduction : to 3' end                                        ||
#> ||                                                                            ||
#> \\============================================================================//
#> 
#> //================================= Running ==================================\\
#> ||                                                                            ||
#> || Load annotation file .Rsubread_UserProvidedAnnotation_pid1296978 ...       ||
#> ||    Features : 13102                                                        ||
#> ||    Meta-features : 13102                                                   ||
#> ||    Chromosomes/contigs : 1                                                 ||
#> ||                                                                            ||
#> || Process BAM file pnp1_1_1_ChrC_71950_78500.bam...                          ||
#> ||    Strand specific : stranded                                              ||
#> ||    Single-end reads are included.                                          ||
#> ||    Total alignments : 24260                                                ||
#> ||    Successfully assigned alignments : 24249 (100.0%)                       ||
#> ||    Running time : 0.00 minutes                                             ||
#> ||                                                                            ||
#> || Process BAM file pnp1_1_2_ChrC_71950_78500.bam...                          ||
#> ||    Strand specific : stranded                                              ||
#> ||    Single-end reads are included.                                          ||
#> ||    Total alignments : 21100                                                ||
#> ||    Successfully assigned alignments : 21098 (100.0%)                       ||
#> ||    Running time : 0.00 minutes                                             ||
#> ||                                                                            ||
#> || Process BAM file wt_1_ChrC_71950_78500.bam...                              ||
#> ||    Strand specific : stranded                                              ||
#> ||    Single-end reads are included.                                          ||
#> ||    Total alignments : 17533                                                ||
#> ||    Successfully assigned alignments : 17531 (100.0%)                       ||
#> ||    Running time : 0.00 minutes                                             ||
#> ||                                                                            ||
#> || Process BAM file wt_2_ChrC_71950_78500.bam...                              ||
#> ||    Strand specific : stranded                                              ||
#> ||    Single-end reads are included.                                          ||
#> ||    Total alignments : 14070                                                ||
#> ||    Successfully assigned alignments : 14066 (100.0%)                       ||
#> ||    Running time : 0.00 minutes                                             ||
#> ||                                                                            ||
#> || Write the final count table.                                               ||
#> || Write the read assignment summary.                                         ||
#> ||                                                                            ||
#> \\============================================================================//
#> 
#> 
#>         ==========     _____ _    _ ____  _____  ______          _____  
#>         =====         / ____| |  | |  _ \|  __ \|  ____|   /\   |  __ \ 
#>           =====      | (___ | |  | | |_) | |__) | |__     /  \  | |  | |
#>             ====      \___ \| |  | |  _ <|  _  /|  __|   / /\ \ | |  | |
#>               ====    ____) | |__| | |_) | | \ \| |____ / ____ \| |__| |
#>         ==========   |_____/ \____/|____/|_|  \_\______/_/    \_\_____/
#>        Rsubread 2.10.4
#> 
#> //========================== featureCounts setting ===========================\\
#> ||                                                                            ||
#> ||             Input files : 4 BAM files                                      ||
#> ||                                                                            ||
#> ||                           pnp1_1_1_ChrC_71950_78500.bam                    ||
#> ||                           pnp1_1_2_ChrC_71950_78500.bam                    ||
#> ||                           wt_1_ChrC_71950_78500.bam                        ||
#> ||                           wt_2_ChrC_71950_78500.bam                        ||
#> ||                                                                            ||
#> ||              Paired-end : no                                               ||
#> ||        Count read pairs : no                                               ||
#> ||              Annotation : R data.frame                                     ||
#> ||      Dir for temp files : .                                                ||
#> ||                 Threads : 1                                                ||
#> ||                   Level : meta-feature level                               ||
#> ||      Multimapping reads : counted                                          ||
#> || Multi-overlapping reads : not counted                                      ||
#> ||   Min overlapping bases : 1                                                ||
#> ||          Read reduction : to 5' end                                        ||
#> ||                                                                            ||
#> \\============================================================================//
#> 
#> //================================= Running ==================================\\
#> ||                                                                            ||
#> || Load annotation file .Rsubread_UserProvidedAnnotation_pid1296978 ...       ||
#> ||    Features : 13102                                                        ||
#> ||    Meta-features : 13102                                                   ||
#> ||    Chromosomes/contigs : 1                                                 ||
#> ||                                                                            ||
#> || Process BAM file pnp1_1_1_ChrC_71950_78500.bam...                          ||
#> ||    Strand specific : stranded                                              ||
#> ||    Single-end reads are included.                                          ||
#> ||    Total alignments : 24260                                                ||
#> ||    Successfully assigned alignments : 24251 (100.0%)                       ||
#> ||    Running time : 0.00 minutes                                             ||
#> ||                                                                            ||
#> || Process BAM file pnp1_1_2_ChrC_71950_78500.bam...                          ||
#> ||    Strand specific : stranded                                              ||
#> ||    Single-end reads are included.                                          ||
#> ||    Total alignments : 21100                                                ||
#> ||    Successfully assigned alignments : 21088 (99.9%)                        ||
#> ||    Running time : 0.00 minutes                                             ||
#> ||                                                                            ||
#> || Process BAM file wt_1_ChrC_71950_78500.bam...                              ||
#> ||    Strand specific : stranded                                              ||
#> ||    Single-end reads are included.                                          ||
#> ||    Total alignments : 17533                                                ||
#> ||    Successfully assigned alignments : 17520 (99.9%)                        ||
#> ||    Running time : 0.00 minutes                                             ||
#> ||                                                                            ||
#> || Process BAM file wt_2_ChrC_71950_78500.bam...                              ||
#> ||    Strand specific : stranded                                              ||
#> ||    Single-end reads are included.                                          ||
#> ||    Total alignments : 14070                                                ||
#> ||    Successfully assigned alignments : 14060 (99.9%)                        ||
#> ||    Running time : 0.00 minutes                                             ||
#> ||                                                                            ||
#> || Write the final count table.                                               ||
#> || Write the read assignment summary.                                         ||
#> ||                                                                            ||
#> \\============================================================================//
#> 
#> building log2-FC per-base ...
#> finished in 0.486620426177979 secs

# Summarize the differential landscape.
SExp <- segmentation(
  data                   = data, 
  nbThreadsFeatureCounts = 1,
  outputDirectory        = working_directory
)
#> changepoint detection on ChrC 71950 78500 plus ...
#>   > model selection ...
#>   > report available in ./DIFFSEGR_TEST/all_models_plus.pdf
#> changepoint detection on ChrC 71950 78500 minus ...
#>   > model selection ...
#>   > report available in ./DIFFSEGR_TEST/all_models_minus.pdf
#> finished in 0.181750535964966 secs
#> counting ...
#> 
#>         ==========     _____ _    _ ____  _____  ______          _____  
#>         =====         / ____| |  | |  _ \|  __ \|  ____|   /\   |  __ \ 
#>           =====      | (___ | |  | | |_) | |__) | |__     /  \  | |  | |
#>             ====      \___ \| |  | |  _ <|  _  /|  __|   / /\ \ | |  | |
#>               ====    ____) | |__| | |_) | | \ \| |____ / ____ \| |__| |
#>         ==========   |_____/ \____/|____/|_|  \_\______/_/    \_\_____/
#>        Rsubread 2.10.4
#> 
#> //========================== featureCounts setting ===========================\\
#> ||                                                                            ||
#> ||             Input files : 4 BAM files                                      ||
#> ||                                                                            ||
#> ||                           pnp1_1_1_ChrC_71950_78500.bam                    ||
#> ||                           pnp1_1_2_ChrC_71950_78500.bam                    ||
#> ||                           wt_1_ChrC_71950_78500.bam                        ||
#> ||                           wt_2_ChrC_71950_78500.bam                        ||
#> ||                                                                            ||
#> ||              Paired-end : no                                               ||
#> ||        Count read pairs : no                                               ||
#> ||              Annotation : R data.frame                                     ||
#> ||      Dir for temp files : .                                                ||
#> ||                 Threads : 1                                                ||
#> ||                   Level : meta-feature level                               ||
#> ||      Multimapping reads : counted                                          ||
#> || Multi-overlapping reads : counted                                          ||
#> ||   Min overlapping bases : 1                                                ||
#> ||                                                                            ||
#> \\============================================================================//
#> 
#> //================================= Running ==================================\\
#> ||                                                                            ||
#> || Load annotation file .Rsubread_UserProvidedAnnotation_pid1296978 ...       ||
#> ||    Features : 69                                                           ||
#> ||    Meta-features : 69                                                      ||
#> ||    Chromosomes/contigs : 1                                                 ||
#> ||                                                                            ||
#> || Process BAM file pnp1_1_1_ChrC_71950_78500.bam...                          ||
#> ||    Strand specific : stranded                                              ||
#> ||    Single-end reads are included.                                          ||
#> ||    Total alignments : 24260                                                ||
#> ||    Successfully assigned alignments : 24260 (100.0%)                       ||
#> ||    Running time : 0.00 minutes                                             ||
#> ||                                                                            ||
#> || Process BAM file pnp1_1_2_ChrC_71950_78500.bam...                          ||
#> ||    Strand specific : stranded                                              ||
#> ||    Single-end reads are included.                                          ||
#> ||    Total alignments : 21100                                                ||
#> ||    Successfully assigned alignments : 21100 (100.0%)                       ||
#> ||    Running time : 0.00 minutes                                             ||
#> ||                                                                            ||
#> || Process BAM file wt_1_ChrC_71950_78500.bam...                              ||
#> ||    Strand specific : stranded                                              ||
#> ||    Single-end reads are included.                                          ||
#> ||    Total alignments : 17533                                                ||
#> ||    Successfully assigned alignments : 17533 (100.0%)                       ||
#> ||    Running time : 0.00 minutes                                             ||
#> ||                                                                            ||
#> || Process BAM file wt_2_ChrC_71950_78500.bam...                              ||
#> ||    Strand specific : stranded                                              ||
#> ||    Single-end reads are included.                                          ||
#> ||    Total alignments : 14070                                                ||
#> ||    Successfully assigned alignments : 14070 (100.0%)                       ||
#> ||    Running time : 0.00 minutes                                             ||
#> ||                                                                            ||
#> || Write the final count table.                                               ||
#> || Write the read assignment summary.                                         ||
#> ||                                                                            ||
#> \\============================================================================//
#> 
#> finished in 0.125626564025879 secs

# Differential expression analysis. We control all the resulting 
# p-values using a Benjamini–Hochberg procedure at level 0.01.
dds <- dea(
  data              = data,
  SExp              = SExp,
  design            = ~ condition,
  sizeFactors       = rep(1,4),
  significanceLevel = 1e-2
)
#>   > differential expression analysis ...
#> renaming the first element in assays to 'counts'
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#> final dispersion estimates
#>   >  multiple testing correction ...
#>   >  normalize denoised log2-FC and coverage ...

# first to fifth regions 
print(SummarizedExperiment::mcols(dds)[1:5,])
#> DataFrame with 5 rows and 37 columns
#>                       modelStart  modelEnd modelMean             featureId
#>                        <numeric> <numeric> <numeric>           <character>
#> ChrC_71950_72255_plus          1       306 0.0257894 ChrC_71950_72255_plus
#> ChrC_72256_72402_plus        307       453 0.6352520 ChrC_72256_72402_plus
#> ChrC_72403_74038_plus        454      2089 0.2826332 ChrC_72403_74038_plus
#> ChrC_74039_74305_plus       2090      2356 0.0910074 ChrC_74039_74305_plus
#> ChrC_74306_74841_plus       2357      2892 0.3207993 ChrC_74306_74841_plus
#>                        baseMean    baseVar   allZero dispGeneEst dispGeneIter
#>                       <numeric>  <numeric> <logical>   <numeric>    <numeric>
#> ChrC_71950_72255_plus    112.50    1085.67     FALSE  0.00000001            5
#> ChrC_72256_72402_plus    483.25   40476.25     FALSE  0.00000001            2
#> ChrC_72403_74038_plus  11037.25 4114258.92     FALSE  0.01999925            2
#> ChrC_74039_74305_plus    122.25     696.25     FALSE  0.00000001           14
#> ChrC_74306_74841_plus   1785.00  236376.67     FALSE  0.02248815            2
#>                         dispFit dispersion  dispIter dispOutlier   dispMAP
#>                       <numeric>  <numeric> <integer>   <logical> <numeric>
#> ChrC_71950_72255_plus 0.0438822  0.0359269         6       FALSE 0.0359269
#> ChrC_72256_72402_plus 0.0230099  0.0189383         6       FALSE 0.0189383
#> ChrC_72403_74038_plus 0.0169537  0.0175717         9       FALSE 0.0175717
#> ChrC_74039_74305_plus 0.0417124  0.0352381         6       FALSE 0.0352381
#> ChrC_74306_74841_plus 0.0183910  0.0192489         9       FALSE 0.0192489
#>                       Intercept condition_pnp1_1_vs_wt SE_Intercept
#>                       <numeric>              <numeric>    <numeric>
#> ChrC_71950_72255_plus   6.39232               0.747234     0.223109
#> ChrC_72256_72402_plus   8.27380               1.085954     0.151892
#> ChrC_72403_74038_plus  13.22995               0.375882     0.135628
#> ChrC_74039_74305_plus   6.65821               0.506696     0.216738
#> ChrC_74306_74841_plus  10.45481               0.626337     0.144130
#>                       SE_condition_pnp1_1_vs_wt WaldStatistic_Intercept
#>                                       <numeric>               <numeric>
#> ChrC_71950_72255_plus                  0.307485                 28.6511
#> ChrC_72256_72402_plus                  0.210628                 54.4714
#> ChrC_72403_74038_plus                  0.191742                 97.5460
#> ChrC_74039_74305_plus                  0.301495                 30.7200
#> ChrC_74306_74841_plus                  0.203189                 72.5374
#>                       WaldStatistic_condition_pnp1_1_vs_wt WaldPvalue_Intercept
#>                                                  <numeric>            <numeric>
#> ChrC_71950_72255_plus                              2.43015         1.55327e-180
#> ChrC_72256_72402_plus                              5.15579          0.00000e+00
#> ChrC_72403_74038_plus                              1.96035          0.00000e+00
#> ChrC_74039_74305_plus                              1.68061         3.07377e-207
#> ChrC_74306_74841_plus                              3.08253          0.00000e+00
#>                       WaldPvalue_condition_pnp1_1_vs_wt  betaConv  betaIter
#>                                               <numeric> <logical> <numeric>
#> ChrC_71950_72255_plus                       1.50926e-02      TRUE         2
#> ChrC_72256_72402_plus                       2.52561e-07      TRUE         2
#> ChrC_72403_74038_plus                       4.99545e-02      TRUE         2
#> ChrC_74039_74305_plus                       9.28388e-02      TRUE         2
#> ChrC_74306_74841_plus                       2.05246e-03      TRUE         2
#>                        deviance  maxCooks  baseMean log2FoldChange     lfcSE
#>                       <numeric> <logical> <numeric>      <numeric> <numeric>
#> ChrC_71950_72255_plus   32.5318        NA    112.50       0.747234  0.307485
#> ChrC_72256_72402_plus   41.1243        NA    483.25       1.085954  0.210628
#> ChrC_72403_74038_plus   67.8832        NA  11037.25       0.375882  0.191742
#> ChrC_74039_74305_plus   33.5075        NA    122.25       0.506696  0.301495
#> ChrC_74306_74841_plus   53.7430        NA   1785.00       0.626337  0.203189
#>                            stat      pvalue        padj log2FoldChangeMean
#>                       <numeric>   <numeric>   <numeric>          <numeric>
#> ChrC_71950_72255_plus   2.43015 1.50926e-02 2.36679e-02          0.0257894
#> ChrC_72256_72402_plus   5.15579 2.52561e-07 2.48953e-06          0.6352520
#> ChrC_72403_74038_plus   1.96035 4.99545e-02 7.33374e-02          0.2826332
#> ChrC_74039_74305_plus   1.68061 9.28388e-02 1.23190e-01          0.0910074
#> ChrC_74306_74841_plus   3.08253 2.05246e-03 3.72683e-03          0.3207993
#>                       log2RefMean log2OtherMean rejectedHypotheses     group
#>                         <numeric>     <numeric>          <logical> <numeric>
#> ChrC_71950_72255_plus    0.185676      0.211465              FALSE         0
#> ChrC_72256_72402_plus    0.877737      1.512989               TRUE         1
#> ChrC_72403_74038_plus    1.849847      2.132480              FALSE         2
#> ChrC_74039_74305_plus    0.192321      0.283328              FALSE         2
#> ChrC_74306_74841_plus    1.037875      1.358675               TRUE         3

# delete working directory 
unlink(working_directory, recursive = TRUE)