Define a New Experiment

loadData allows for loading various pieces of information related to RNA-Seq samples and the loci that will be analyzed.

Usage

newExperiment(
  sampleInfo,
  referenceCondition,
  otherCondition,
  loci,
  nbThreads = 1,
  nbThreadsByLocus = 1
)

Arguments

sampleInfo

A String. Path to the file with sample information. The file includes the following columns:

sample : identifier for the sample ;

condition : identifier for the condition ;

replicate : identifier for the replicate ;

bam : path to the bam file ;

coverage : path to the coverage file (in rds format);

isPairedEnd : logical indicating if the library contain paired-end reads or not (TRUE/FALSE);

strandSpecific : integer indicating if a strand-specific analysis should be performed. ‘0’ (unstranded), ‘1’ (stranded) and ‘2’ (reversely stranded). WARNING: All samples must be consistently stranded or non-stranded.

Alternatively, one may provide a Data.frame that contains the same columns.

referenceCondition

A String. The reference condition of the RNA-Seq experiment.

otherCondition

A String. The alternative condition that will be compare to the reference one.

loci

A String. Path to the file with loci information. The file includes the following columns:

seqid : chromosome identifier (match one of the chromosome in the bam) for the target genomic region ;

chromStart : start position for the target genomic region ;

chromEnd : end position for the target genomic region ;

locusID : unique id for the target genomic region.

Alternatively, one may provide a Data.frame that contains the same columns.

nbThreads

An Integer. The overall number of threads used for the analysis.

nbThreadsByLocus

An Integer. The number of threads dedicated to each locus. nbThreadsByLocus cannot be greater than nbThreads. The number of loci analyzed in parallel is equal to the divisor of nbThreads by nbThreadsByLocus.

Value

A List. The list containing loaded data.